+++ CHANGELOG.md
@@ -16,5 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CameraMouseWobble
 - Task
 - Tween
+- ShowWhenAttribute
+- Tween.Custom
 
 ### Added